refactor: Player가 input에 대한 검증을 하도록 수정
